Do CloudSearch API migration - lucene parser of cloudsearch
Reddit follows lucene query style. This required to convert given query
to cloudsearch query style because cloudsearch API 2011 doesn't recognize
lucene syntax. But cloudsearch API 2013 does! What we have to do is to
indicate parser 'lucene'.

So now it is not required to convert the query style. l2cs module is not
used anymore.