Use searchresultsubreddit template for rollbacked subreddit search result
When there is no search results of subreddit but is some link results,
this codes shows subreddit facets of the result links.
But it is using wrong template. This patch make it use the same
template with subreddit results.

I'm not sure why this doesn't happen in the Reddit site.